refactor: extract quality_helpers.py and add aggregate health Prometheus gauges
Move _detect_gaps and _count_weekdays_between from data_quality.py into a
shared quality_helpers.py so the lightweight nightly staleness sweep can
reuse them without importing the full DataQualityService.

Add markethawk_aggregate_staleness_hours and markethawk_aggregate_gap_days
Prometheus gauges (per-universe, multiprocess_mode=livemax) to core/metrics.py.
